Changed to a 8-bit architecture.

+++ src/CSU/include/CSU.hpp
@@ -12,10 +12,10 @@ class CSU{
 
         void printState(uint64_t size = 16);
 
-        void setMemoryRegion(uint64_t start, std::vector<int64_t> newValues);
+        void setMemoryRegion(uint64_t start, std::vector<int8_t> newValues);
     
     private:
-        std::vector<int64_t> memory;
-        uint64_t pc = 0;
-        uint64_t lastPc = 0;
+        std::vector<int8_t> memory;
+        uint16_t pc = 0;
+        uint16_t lastPc = 0;
 };
